backend/pyproject.toml
[project]
name = "dydat-backend"
version = "0.1.0"
description = "Dydat — Adaptive AI math tutor backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "anthropic>=0.43,<1",
    "python-jose[cryptography]>=3.3,<4",
    "bcrypt>=4.0,<6",
    "sse-starlette>=2.0,<3",
    "networkx>=3.4,<4",
    "email-validator>=2.0,<3",
    "httpx>=0.28,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"app/llm/prompts/system_prompt.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: test con LLM reale (richiede ANTHROPIC_API_KEY, eseguire con --run-integration)",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
